+++ js/src/editor.js
@@ -1,0 +1,7 @@
+/**
+ * @file editor.js
+ * @description Block editor customizations and core block configurations
+ */
+
+// Core block configurations
+import './core-blocks/media-text';